is-project/requirements.txt
africastalking==1.2.7
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.7
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
django-rest-knox==4.2.0
djangorestframework==3.15.1
filelock==3.14.0
idna==3.7
pillow==10.4.0
platformdirs==4.2.2
pycparser==2.22
reportlab==4.2.2
requests==2.32.3
schema==0.7.7
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.2
